requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
DateTime==5.5
distlib==0.3.7
Django==5.0.6
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
env==0.1.0
filelock==3.12.3
pillow==10.3.0
platformdirs==3.10.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.24.5
zope.interface==6.4.post2
